refactor: extract resident needs policy
