+++ v2/orchestrator-rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cosyworld-orchestrator"
-version = "0.0.206"
+version = "0.0.207"
 edition = "2021"
 publish = false
