+++ v2/orchestrator-rust/src/residents/mod.rs
@@ -1,3 +1,4 @@
 mod continuity;
+mod needs;
 
 pub(super) use continuity::*;
